scores are floats

+++ infrastructure/db/exam_template_schema.py
@@ -1,6 +1,6 @@
 from infrastructure.db.database import Base
 from sqlalchemy.orm import relationship
-from sqlalchemy import Column, Integer, String, Boolean
+from sqlalchemy import Column, Float, String, Boolean
 from sqlalchemy.dialects.postgresql import UUID
 import uuid
 
@@ -15,7 +15,7 @@ class ExamTemplate(Base):
     name = Column(String(50), nullable=False)
     course_id = Column(UUID(as_uuid=True), nullable=False)
     state = Column(String(10), default="Draft")  # Enum(ExamStateEnum))
-    max_score = Column(Integer, default=10)
+    max_score = Column(Float, default=10)
     has_multiple_choice = Column(Boolean, default=True)
     has_written = Column(Boolean, default=False)
     has_media = Column(Boolean, default=False)
